==> 02_Workbook_Cleansing.ipynb <==
## Yellow Taxi Tripdata
### Bereinigung und Vorbereitung

#### Data Cleansing (Datenbereinigung)
Data Cleansing konzentriert sich darauf, schlechte Daten zu finden und zu beheben. Es geht um die Qualität der Daten. Stell dir vor, du korrigierst einen Text auf Rechtschreibfehler, fehlende Wörter oder doppelte Sätze.

Typische Aufgaben sind:

- Fehlende Werte auffüllen oder entfernen.

- Doppelte Datensätze löschen.

- Inkonsistente Einträge korrigieren (z. B. "USA" und "Vereinigte Staaten" zu standardisieren).

- Ausreißer und unsinnige Werte entfernen.

#### Data Wrangling (Datenaufbereitung)
Data Wrangling ist der übergeordnete, umfassende Prozess, um Rohdaten in eine nutzbare Form für die Analyse zu bringen. Es ist wie das gesamte Schreiben eines Berichts, nicht nur das Korrigieren. Es beinhaltet das Cleansing, geht aber noch weit darüber hinaus.

Typische Aufgaben sind:

- Das Cleansing der Daten.

- Das Verknüpfen mehrerer Datensätze (wie wir mit dem lookup_df machen werden).

- Das Umstrukturieren der Daten (z. B. Spalten in Zeilen umwandeln).

- Das Anreichern der Daten (z. B. neue Spalten für "Tag der Woche" erstellen).

**Zusammenfassend**: Das Ziel von Cleansing ist es, die Daten sauber und korrekt zu machen. Das Ziel von Wrangling ist es, die sauberen Daten bereit für die Analyse zu machen.

Zuerst widmen wir uns der Bereinigung des großen DataFrames. Wir korrigieren die Datentypen der Spalten, insbesondere der Datums- und Zeitangaben, kümmern uns um fehlende Werte und filtern offensichtliche Fehler wie Fahrten mit 0 Dollar Preis oder 0 Minuten Dauer heraus. Ziel ist ein sauberer, zuverlässiger Datensatz.

import pandas as pd
#import pyarrow
from fastparquet import ParquetFile

# Einlesen des Parquet-Files
pf = ParquetFile('NYTaxi-TripData/yellow_tripdata.parquet')
df = pf.to_pandas()

df.head()

df.info()

### Legende
#### Bevor wir so richtig loslegen hier noch eine Übersicht über die Bedeutung der Spalten:
##### Die Einträge in einigen Spalten referenzieren Werte die auch nachfolgend angegeben sind.
- `VendorID`: Eine ID des Taxi-Anbieters (z. B. 1 für Creative Mobile Technologies, 2 für Curb Mobility).
    - 1: Creative Mobile Technologies, LLC
    - 2: VeriFone Inc.

- `tpep_pickup_datetime`: Das Datum und die Uhrzeit, wann die Taxifahrt begonnen hat.

- `tpep_dropoff_datetime`: Das Datum und die Uhrzeit, wann die Fahrt beendet wurde.

- `passenger_count`: Die Anzahl der Passagiere auf der Fahrt.

- `trip_distance`: Die zurückgelegte Entfernung in Meilen.

- `PULocationID`: Die ID der Zone, in der die Fahrt begonnen hat. (Zone-Lookup-Frame)

- `DOLocationID`: Die ID der Zone, in der die Fahrt geendet hat.  (Zone-Lookup-Frame)

- `RatecodeID`: Ein Code für den Tarif der Fahrt (z. B. 1 = Standard, 2 = Fahrt zum JFK-Flughafen).
    - 1: Standard rate
    - 2: JFK
    - 3: Newark
    - 4: Nassau or Westchester
    - 5: Negotiated fare
    - 6: Group ride
- `payment_type`: Der Zahlungs-Typ (z. B. 1 = Kreditkarte, 2 = Barzahlung).
    - 1: Credit card
    - 2: Cash
    - 3: No charge
    - 4: Dispute
    - 5: Unknown
    - 6: Voided trip

- `fare_amount`: Der reine Fahrpreis (ohne Extras).

- `congestion_surcharge`: Eine eventuell erhobene Staugebühr.

#### store_and_fwd_flag
Hier noch eine Erklärung zu der Spalte store_and_fwd_flag:
Das `store_and_fwd_flag` sagt aus, ob die Daten einer Taxifahrt sofort an den Server des Anbieters gesendet werden konnten oder nicht.

Wenn das Signal im Taxi schlecht war, wurden die Daten lokal auf dem Taxameter gespeichert (store) und erst später, als die Verbindung wieder da war, weitergeleitet (forward).
- 'Y': Die Daten wurden erst später weitergeleitet.
- 'N': Die Daten wurden sofort gesendet.

Für eine erste, einfache Analyse ist der store_and_fwd_flag vielleicht nicht das wichtigste Feature.

Aber es kann wichtig sein, um die Datenqualität zu beurteilen. Wenn viele Fahrten dieses Flag haben, könnte das auf ein schlechtes GPS-Signal hindeuten, was wiederum die Genauigkeit der Fahrzeiten und der Entfernungen beeinflussen kann. Das wäre dann ein wichtiger Punkt, den man bei der Interpretation der Ergebnisse berücksichtigen müsste.

Auch könnte es sein, dass es für Machine-Learning-Modelle wichtig ist. Vielleicht gibt es einen Zusammenhang zwischen dem Flag und der Höhe des Trinkgeldes oder der Fahrtdauer. Aber das sind eher fortgeschrittene Themen.

Kurz gesagt: Es ist ein kleiner, aber feiner Hinweis auf die Umstände, unter denen die Daten gesammelt wurden.

#### Datentypen korrigieren
Der erste Schritt ist, die Datentypen der Datums- und Zeitspalten korrekt einzustellen. Wie wir besprochen haben, müssen die Spalten tpep_pickup_datetime und tpep_dropoff_datetime in das datetime-Format umgewandelt werden.

df['tpep_pickup_datetime'] = pd.to_datetime(df['tpep_pickup_datetime'])
df['tpep_dropoff_datetime'] = pd.to_datetime(df['tpep_dropoff_datetime'])
df.info()

# df.isnull().sum <- ergibt eine unübersichliche Ausgabe

# wir sind nur an Spalten interessiert, die mindestens 1 fehlenden Wert haben.
df.isnull().sum()[df.isnull().sum() > 0]

Für die `congestion_surcharge`Wert gehen wir davon aus, dass für die Fahrt keine Staugebühr erhoben wurde.
Wir ersetzen die fehlenden Werte (NaN) durch eine 0

# Ersetze alle fehlenden Werte in der congestion_surcharge-Spalte durch 0
#df['congestion_surcharge'].fillna(0, inplace=True) <-- erzeugt eine Warnung

# So geht es ohne Warnmeldung
df['congestion_surcharge'] = df['congestion_surcharge'].fillna(0)

# Nochmal die Ausgabe zur Kontrolle
df.isnull().sum()[df.isnull().sum() > 0]


#### Interpretation der Fehlwerte
Die Tatsache, dass alle fünf Spalten exakt die gleiche Anzahl an fehlenden Werten haben, ist ein sehr starkes Signal. Es ist höchst unwahrscheinlich, dass das Zufall ist.

##### Interpretation:
Diese 527.484 fehlenden Werte betreffen wahrscheinlich die gleichen Fahrten. Es scheint, als wären für diese Trips aus irgendeinem Grund die Daten für alle diese Spalten auf einmal nicht erfasst worden.

##### Nächster Schritt:
Da uns für diese Fahrten so viele wichtige Informationen fehlen, sind die Zeilen für unsere Analyse wahrscheinlich nicht nutzbar. Die beste Methode, um damit umzugehen, ist, diese Zeilen einfach aus dem Datensatz zu entfernen.

# (vorher)
df.shape

df = df.dropna(subset=['VendorID', 'passenger_count', 'RatecodeID', 'store_and_fwd_flag', 'payment_type'])

# (nachher) - Die Differenz sollte wieder 527.484 ergeben...
df.shape

- Die Methode `dropna()` entfernt Zeilen mit fehlenden Werten (NaN, None, etc.)
- Der Parameter `subset` gibt an, dass nur die genannten Spalten auf fehlende Werte überprüft werden sollen
- Es werden nur Zeilen entfernt, die in mindestens einer der folgenden Spalten fehlende Werte haben:
    - 'VendorID'
    - 'passenger_count'
    - 'RatecodeID'
    - 'store_and_fwd_flag'
    - 'payment_type'

- Das Ergebnis wird dem ursprünglichen DataFrame `df` zugewiesen, was bedeutet, dass der originale DataFrame überschrieben wird

Diese Operation ist typisch für die Datenbereinigung, wobei unvollständige Datensätze entfernt werden, bevor weitere Analysen durchgeführt werden. Alterntiv können wir natürlich auch fehlende Daten einfach ersetzten - dazu werden wir später noch mehr erfahren.




#### Filtern unnötiger Datensätze
Wir filtern den Datensatz so, dass nur noch Fahrten übrig bleiben, bei denen sowohl die Entfernung als auch der Fahrpreis größer als 0 sind.

# Filtert den DataFrame, um nur Fahrten mit trip_distance und fare_amount größer 0 zu behalten
df = df[(df['trip_distance'] > 0) & (df['fare_amount'] > 0)]
df.shape

### Speichern & Feierabend
Natürlich wollen wir unseren schönen, sauberen und bereinigten Datensatz für zukünftige Missionen speichern!

# Speichern des aktuellen Dataframes
df.to_parquet('NYTaxi-TripData/new_york_taxi_cleaned.parquet', index=False)
