==> 04_Workbook_Analytic.ipynb <==
### Easy Analytics (before breakfast ;-)
Wir laden unseren finalen Datensatz und legen los...

import pandas as pd
from fastparquet import ParquetFile

# Load Data
pf = ParquetFile('NYTaxi-TripData/new_york_taxi_cleaned_final.parquet')
df = pf.to_pandas()
df.head()

---

1. #### Wie viele Fahrten gibt es insgesamt?
* Aufgabe: Finde die Anzahl der Zeilen (Fahrten) im DataFrame.

df.shape[0]

#### 2. Wie viele Spalten gibt es im Datensatz?
* Aufgabe: Finde die Anzahl der Spalten im DataFrame.

df.shape[1]

#### 3. Wie viele einzigartige Anbieter (VendorID) gibt es?
* Aufgabe: Zähle, wie viele verschiedene Taxi-Anbieter im Datensatz vorkommen.

df['VendorID'].nunique()

#### 4. Was ist die maximale Anzahl von Passagieren in einer Fahrt?
* Aufgabe: Finde den höchsten Wert in der Spalte passenger_count.

df['passenger_count'].max()

#### 5. Was ist der niedrigste Fahrpreis (fare_amount)?
* Aufgabe: Finde den niedrigsten Wert in der Spalte fare_amount.
* Aufgabe: Finde die Anzahl der Fahrten mit einem Fahrpreis unter $2.50
* Aufgabe: Gib das Ergebnis in Prozent aus

df['fare_amount'].min()

# Anzahl der Fahrten mit Fahrpreis unter 2,5 ermitteln
anzahl_niedrige_preise = df[df['fare_amount'] < 2.5].shape[0]

print(f"Es gibt {anzahl_niedrige_preise} Fahrten mit einem Fahrpreis unter 2,5 $")

# Optional: Prozentsatz dieser Fahrten an der Gesamtmenge
prozentsatz = (anzahl_niedrige_preise / df.shape[0]) * 100
print(f"Das entspricht {prozentsatz:.2f}% aller Fahrten im Datensatz")


#### 6. Was ist der durchschnittliche Trinkgeldbetrag (tip_amount)?
* Aufgabe: Berechne den Mittelwert der Trinkgelder.

df['tip_amount'].mean()
